RenderTargetの生成方法など変更

+++ MyLibrary/Camera.cpp
@@ -31,7 +31,7 @@ DirectX::XMMATRIX Camera::GetViewAndProjectionMat()
 	DirectX::XMMATRIX mapProjection = DirectX::XMMatrixPerspectiveFovLH
 	(
 		DirectX::XMConvertToRadians(fovY), //画角
-		Library::GetWindowWidth() / Library::GetWindowHeight(),//アスペクト比
+		(float)Library::GetWindowWidth() / (float)Library::GetWindowHeight(),//アスペクト比
 		nearNum, farNum//描画範囲
 	);
